Creacion y validacion LibroService

+++ src/main/java/com/biblioteca/bibliotecacrud/repository/ILibroRepo.java
@@ -3,7 +3,7 @@ package com.biblioteca.bibliotecacrud.repository;
 import com.biblioteca.bibliotecacrud.entity.Libro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ILibroRepo extends JpaRepository<Libro, Integer> {
+public interface ILibroRepo extends JpaRepository<Libro, Long> {
     //Verifica si hay un libro con el mismo titulo
     boolean existsByLibro(String Libro);
 }
